requirements.txt
# --- Framework principal ---
Flask==3.0.3

# --- ORM et Base de données ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
PyMySQL==1.1.1  # pour se connecter à MySQL (Railway)
cryptography==43.0.1  # nécessaire pour Railway auth

# --- Authentification et Sécurité ---
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Flask-Limiter==3.8.0

# --- Gestion des variables d’environnement ---
python-dotenv==1.0.1

# --- Outils divers utilisés ---
WTForms==3.1.2
email-validator==2.2.0
requests==2.32.3
pandas==2.2.3  # utile pour la migration
gunicorn==23.0.0  # serveur de production pour Render

# --- Si tu utilises le modèle IA ---
scikit-learn==1.5.2
numpy==2.1.3
